speed up clusters page by making it a server component
